Compat SPE - Remove hard hearing dependency (#10652)
