+++ addons/compat_spe/CfgWeapons.hpp
@@ -1,5 +1,4 @@
 class CfgWeapons {
-    #include "CfgWeapons\helmets.hpp"
     #include "CfgWeapons\launchers.hpp"
     #include "CfgWeapons\weapons.hpp"
 };
