+++ addons/compat_spe/compat_spe_hearing/CfgWeapons.hpp
@@ -1,0 +1,145 @@
+class CfgWeapons {
+    class H_SPE_HelmetB;
+    class H_SPE_Hat: H_SPE_HelmetB {};
+
+    //ALLIED PILOT
+    class H_SPE_US_Helmet_Pilot: H_SPE_Hat {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_US_Helmet_Pilot_Op: H_SPE_US_Helmet_Pilot {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_US_Helmet_Pilot_Glasses_Up: H_SPE_US_Helmet_Pilot {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_US_Helmet_Pilot_Glasses_Down: H_SPE_US_Helmet_Pilot {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_US_Helmet_Pilot_Respirator: H_SPE_US_Helmet_Pilot {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_US_Helmet_Pilot_Respirator_Glasses_Up: H_SPE_US_Helmet_Pilot_Respirator {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_US_Helmet_Pilot_Respirator_Glasses_Down: H_SPE_US_Helmet_Pilot_Respirator {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_US_Helmet_Pilot_SWDG_Respirator: H_SPE_US_Helmet_Pilot {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_US_Helmet_Pilot_SWDG_low_Respirator: H_SPE_US_Helmet_Pilot {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_US_Helmet_H15: H_SPE_US_Helmet_Pilot {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_US_Helmet_H15_Op: H_SPE_US_Helmet_H15 {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_US_Helmet_H15_SWDG: H_SPE_US_Helmet_H15 {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_US_Helmet_H15_SWDG_low: H_SPE_US_Helmet_H15 {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_US_Helmet_H15_O2: H_SPE_US_Helmet_H15 {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_US_Helmet_H15_B7_O2: H_SPE_US_Helmet_H15_O2 {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_US_Helmet_H15_B7_low_O2: H_SPE_US_Helmet_H15_O2 {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_US_Helmet_H15_SWDG_O2: H_SPE_US_Helmet_H15_O2 {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_US_Helmet_H15_SWDG_low_O2: H_SPE_US_Helmet_H15_O2 {
+        HEARING_PROTECTION_EARMUFF;
+    };
+
+    //ALLIED TANKER
+    class H_SPE_US_Helmet_Tank: H_SPE_HelmetB {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_US_Helmet_Tank_NG: H_SPE_US_Helmet_Tank {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_US_Helmet_Tank_Nun: H_SPE_US_Helmet_Tank {
+        HEARING_PROTECTION_OPEN;
+    };
+    class H_SPE_US_Helmet_Tank_M1_OS: H_SPE_US_Helmet_Tank {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_US_Helmet_Tank_M1_NS: H_SPE_US_Helmet_Tank_M1_OS {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_US_Helmet_Tank_M1_Scrim: H_SPE_US_Helmet_Tank_M1_OS {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_US_Helmet_Tank_polar: H_SPE_US_Helmet_Tank {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_US_Helmet_Tank_polar_tapes: H_SPE_US_Helmet_Tank_polar {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_US_Helmet_Tank_polar_low: H_SPE_US_Helmet_Tank_polar {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_US_Helmet_Tank_tapes: H_SPE_US_Helmet_Tank {
+        HEARING_PROTECTION_EARMUFF;
+    };
+
+    //AXIS PILOT
+    class H_SPE_GER_LW_PilotHelmet_base: H_SPE_Hat {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_GER_LW_PilotHelmet: H_SPE_GER_LW_PilotHelmet_base {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_GER_LW_S_PilotHelmet: H_SPE_GER_LW_PilotHelmet_base {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_GER_LW_PilotHelmet_Mask_base: H_SPE_Hat {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_GER_LW_PilotHelmet_Mask: H_SPE_GER_LW_PilotHelmet_Mask_base {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_GER_LW_PilotHelmet_S_Mask: H_SPE_GER_LW_PilotHelmet_Mask_base {
+        HEARING_PROTECTION_EARMUFF;
+    };
+
+    //AXIS TANKER
+    class H_SPE_GER_TankPrivateCap: H_SPE_Hat {
+        HEARING_PROTECTION_OPEN;
+    };
+    class H_SPE_GER_TankOfficerCap: H_SPE_Hat {
+        HEARING_PROTECTION_OPEN;
+    };
+    class H_SPE_GER_Fieldcap: H_SPE_Hat {
+        HEARING_PROTECTION_OPEN;
+    };
+    class H_SPE_GER_Fieldcap2: H_SPE_GER_Fieldcap {
+        HEARING_PROTECTION_OPEN;
+    };
+    class H_SPE_GER_TankPrivateCap2: H_SPE_GER_TankPrivateCap {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_GER_TankOffzCap2: H_SPE_GER_TankPrivateCap {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_GER_TankOfficerCap2: H_SPE_GER_TankOfficerCap {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_GER_SPGPrivateCap: H_SPE_Hat {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_GER_ST_Tanker_Cap2: H_SPE_GER_Fieldcap2 {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class H_SPE_GER_ST_Tanker_Cap: H_SPE_GER_Fieldcap {
+        HEARING_PROTECTION_EARMUFF;
+    };
+
+};
